+++ packages/cli/src/util/telemetry/commands/alias/index.ts
@@ -1,0 +1,24 @@
+import { TelemetryClient } from '../..';
+
+export class AliasTelemetryClient extends TelemetryClient {
+  trackCliSubcommandLs(actual: string) {
+    this.trackCliSubcommand({
+      subcommand: 'ls',
+      value: actual,
+    });
+  }
+
+  trackCliSubcommandSet(actual: string) {
+    this.trackCliSubcommand({
+      subcommand: 'set',
+      value: actual,
+    });
+  }
+
+  trackCliSubcommandRemove(actual: string) {
+    this.trackCliSubcommand({
+      subcommand: 'rm',
+      value: actual,
+    });
+  }
+}