+++ packages/cli/src/util/telemetry/index.ts
@@ -1,5 +1,7 @@
 import { randomUUID } from 'node:crypto';
 import type { Output } from '../output';
+import os from 'node:os';
+import { GlobalConfig } from '@vercel-internals/types';
 
 const LogLabel = `['telemetry']:`;
 
@@ -14,6 +16,8 @@ interface Options {
 }
 
 interface Event {
+  teamId?: string;
+  sessionId?: string;
   id: string;
   key: string;
   value: string;
@@ -25,6 +29,14 @@ export class TelemetryClient {
   store: TelemetryEventStore;
 
   protected redactedValue = '[REDACTED]';
+  protected redactedArgumentsLength = (args: string[]) => {
+    if (args && args.length === 1) {
+      return 'ONE';
+    } else if (args.length > 1) {
+      return 'MANY';
+    }
+    return 'NONE';
+  };
 
   constructor({ opts }: Args) {
     this.output = opts.output;
@@ -74,9 +86,9 @@ export class TelemetryClient {
     }
   }
 
-  protected trackCliOption(eventData: { flag: string; value: string }) {
+  protected trackCliOption(eventData: { option: string; value: string }) {
     this.track({
-      key: `flag:${eventData.flag}`,
+      key: `option:${eventData.option}`,
       value: eventData.value,
     });
   }
@@ -85,6 +97,59 @@ export class TelemetryClient {
     this.track({
       key: `flag:${flag}`,
       value: 'TRUE',
+    });
+  }
+
+  protected trackCPUs() {
+    this.track({
+      key: 'cpu_count',
+      value: String(os.cpus().length),
+    });
+  }
+
+  protected trackPlatform() {
+    this.track({
+      key: 'platform',
+      value: os.platform(),
+    });
+  }
+
+  protected trackArch() {
+    this.track({
+      key: 'arch',
+      value: os.arch(),
+    });
+  }
+
+  protected trackCI(ciVendorName?: string) {
+    if (ciVendorName) {
+      this.track({
+        key: 'ci',
+        value: ciVendorName,
+      });
+    }
+  }
+
+  protected trackVersion(version?: string) {
+    if (version) {
+      this.track({
+        key: 'version',
+        value: version,
+      });
+    }
+  }
+
+  protected trackDefaultDeploy() {
+    this.track({
+      key: 'default-deploy',
+      value: 'TRUE',
+    });
+  }
+
+  protected trackExtension(extension: string) {
+    this.track({
+      key: 'extension',
+      value: extension,
     });
   }
 
@@ -102,15 +167,32 @@ export class TelemetryEventStore {
   private events: Event[];
   private output: Output;
   private isDebug: boolean;
+  private sessionId: string;
+  private teamId: string = 'NO_TEAM_ID';
+  private config: GlobalConfig['telemetry'];
 
-  constructor(opts: { output: Output; isDebug?: boolean }) {
+  constructor(opts: {
+    output: Output;
+    isDebug?: boolean;
+    config: GlobalConfig['telemetry'];
+  }) {
     this.isDebug = opts.isDebug || false;
     this.output = opts.output;
+    this.sessionId = randomUUID();
     this.events = [];
+    this.config = opts.config;
   }
 
   add(event: Event) {
+    event.sessionId = this.sessionId;
+    event.teamId = this.teamId;
     this.events.push(event);
+  }
+
+  updateTeamId(teamId?: string) {
+    if (teamId) {
+      this.teamId = teamId;
+    }
   }
 
   get readonlyEvents() {
@@ -121,12 +203,28 @@ export class TelemetryEventStore {
     this.events = [];
   }
 
+  get enabled() {
+    if (process.env.VERCEL_TELEMETRY_DISABLED) {
+      return false;
+    }
+
+    return this.config?.enabled === false ? false : true;
+  }
+
   save() {
     if (this.isDebug) {
-      this.output.debug(`${LogLabel} Flushing Events`);
+      // Intentionally not using `this.output.debug` as it will
+      // not write to stderr unless it is run with `--debug`
+      this.output.log(`${LogLabel} Flushing Events`);
       this.events.forEach(event => {
-        this.output.debug(JSON.stringify(event));
+        this.output.log(JSON.stringify(event));
       });
+
+      return;
+    }
+
+    if (this.enabled) {
+      // send events to the server
     }
   }
 }