+++ packages/node/test/unit/dev.test.ts
@@ -10,6 +10,8 @@ import { setTimeout } from 'timers/promises';
 
 vi.setConfig({ testTimeout: 20 * 1000 });
 
+const isWindows = process.platform === 'win32';
+
 const [NODE_MAJOR] = process.versions.node.split('.').map(v => Number(v));
 
 function testForkDevServer(entrypoint: string) {
@@ -63,7 +65,19 @@ async function withDevServer(
   } finally {
     const elapsed = Date.now() - start;
     if (runningTimeout) await setTimeout(runningTimeout - elapsed);
-    child.send('shutdown', error => error && child.kill(9));
+    child.send('shutdown', error => {
+      if (error) {
+        console.log('Shutdown error:', error);
+        try {
+          child.kill(9);
+        } catch (error) {
+          // In Node 22, there's a bug where attempting to kill a child process
+          // results in an EPERM error. Ignore the error in that case.
+          // See: https://github.com/nodejs/node/issues/51766
+          console.log('Ignoring kill error:', error);
+        }
+      }
+    });
     if (child.exitCode === null) await once(child, 'exit');
   }
 }
@@ -105,18 +119,21 @@ async function withDevServer(
         { runningTimeout: 300 }
       ));
 
-    test('with `waitUntil` from context rejecting a promise ', () =>
-      withDevServer(
-        './wait-until-ctx-node-rejected.js',
-        async (url: string) => {
-          const response = await fetch(
-            `${url}/api/wait-until-ctx-node-rejected`
-          );
-          await setTimeout(100); // wait a bit for waitUntil resolution
-          expect(response.status).toBe(200);
-        },
-        { runningTimeout: 100 }
-      ));
+    test.skipIf(isWindows)(
+      'with `waitUntil` from context rejecting a promise ',
+      () =>
+        withDevServer(
+          './wait-until-ctx-node-rejected.js',
+          async (url: string) => {
+            const response = await fetch(
+              `${url}/api/wait-until-ctx-node-rejected`
+            );
+            await setTimeout(100); // wait a bit for waitUntil resolution
+            expect(response.status).toBe(200);
+          },
+          { runningTimeout: 100 }
+        )
+    );
 
     test('exporting GET', () =>
       withDevServer('./web-handlers-node.js', async (url: string) => {